tests/wpt/web-platform-tests/2dcontext/pixel-manipulation/2d.imageData.object.string.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>Canvas test: 2d.imageData.object.string</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/2dcontext/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/2dcontext/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.imageData.object.string</h1>
<p class="desc">ImageData.data converts strings to numbers with ToNumber</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("ImageData.data converts strings to numbers with ToNumber");
_addTest(function(canvas, ctx) {

var imgdata = ctx.getImageData(0, 0, 10, 10);
imgdata.data[0] = 100;
imgdata.data[0] = "110";
_assertSame(imgdata.data[0], 110, "imgdata.data[\""+(0)+"\"]", "110");
imgdata.data[0] = 100;
imgdata.data[0] = "0x78";
_assertSame(imgdata.data[0], 120, "imgdata.data[\""+(0)+"\"]", "120");
imgdata.data[0] = 100;
imgdata.data[0] = " +130e0 ";
_assertSame(imgdata.data[0], 130, "imgdata.data[\""+(0)+"\"]", "130");


});
</script>
